+++ NEWS.md
@@ -1,9 +1,13 @@
 # User visible changes in `AMORS` package
 
+## Version 0.3.2
+
 - `@public` macro to declare non-exported symbols as *public*, a concept introduced in
   Julia 1.11. The macro does nothing on older Julia versions.
 
 - Fix `best_scaling_factor`.
+
+- Make tests reproducible.
 
 - Improve doc.